Add donate button
This commit adds a donate button to the campaign details page.

The button is disabled if the user is not logged in or if the campaign is
already funded.

The button is enabled if the user is logged in and the campaign is not
already funded.

Clicking the button will trigger a donation to the campaign.